Unsynced working example
Everything is cued up and good to go. Now comes the fun part of syncing the lyrics with the song
