+++ assets/js/script.mjs
@@ -1,16 +1,21 @@
 import { default as Player } from "./modules/PlayerManager.mjs";
 
 const play = document.getElementById("play");
-const message = document.getElementById("message");
 
 try {
 	if(typeof BroadcastChannel !== "function") {
 		throw new Error("BroadcastChannel API is not supported");
 	}
-	const player = new Player();
-	play.addEventListener("click",() => player.play());
+
+	const mediaElement = document.getElementById("still-alive");
+
+	const player = new Player(mediaElement);
+	play.addEventListener("click",() => player.init());
 } catch(error) {
+	const message = document.getElementById("message");
+
 	play.classList.add("unsupported");
 	play.innerText = "Your browser can not play this demo";
+
 	message.innerText = error;
 }
