+++ index.html
@@ -15,6 +15,7 @@
 	<footer>
 		<a href="https://github.com/VictorWesterlund/still-alive">Source on GitHub</p>
 	</footer>
+	<audio id="still-alive" src="assets/media/still-alive.webm"></audio>
 	<script type="module" src="assets/js/script.mjs"></script>
 </body>
 </html>
